app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.19.2
fastapi==0.121.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-elasticsearch==0.4.0
langchain-mcp-adapters==0.1.12
langchain-openai==0.3.35
langchain-redis==0.2.4
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.42
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.21.0
ml_dtypes==0.5.3
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
ply==3.11
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-ulid==3.1.0
PyYAML==6.0.3
redis==6.4.0
redisvl==0.11.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
